hodl_backend/requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
aiosqlite
pydantic
requests
aiosqlite
solana
solders
pytest
pytest-asyncio
